Added GPSwaves.py and related test scripts
GPSwaves.py contains the function for GPSwaves. There is test data that was used to test the function stored in GPSwaves_testdata.mat and the output from the test data using the matlab function stored in GPSwaves-testdata-out.mat. The GPSwaves-test.py script takes in the test data, runs the python function and compares to the matlab outputs.