Completes first draft of the simple config.txt
This is the first commit in the process to making a more useful
configuration file and class. The new simple config file only
contains 4 parameters that you can change that are regularly changed.
The rest of the parameters will be stored in the class and can be
changed but generally should not be.